[#14] Code splitting into components
