+++ index.html
@@ -11,6 +11,6 @@
 <body>
   <section class="container" id="container"></section>
   <button class="button button-footer" id="add-lists">Create new list</button>
-  <script src="script.js"></script>
+  <script type="module" src="index.js"></script>
 </body>
 </html>
