+++ state/root.state.js
@@ -1,0 +1,12 @@
+export const notifyStateUpdated = () => {
+  const jsonRootState = JSON.stringify(rootState);
+  localStorage.setItem("data", jsonRootState);
+};
+
+export const restoreState = () => {
+  const strRootArray = localStorage.getItem("data");
+  const parsedRootState = JSON.parse(strRootArray);
+  return parsedRootState;
+};
+
+export const rootState = restoreState() || [];
